pyproject.toml
[project]
name = "gql-book-server"
version = "0.0.1"
description = ""
authors = [
    {name = "Mark Larah",email = "mark@larah.me"}
]
readme = "README.md"
requires-python = "~=3.13.0"
dependencies = [
    "strawberry-graphql[fastapi] (>=0.285.0,<0.286.0)",
    "uvicorn (>=0.38.0,<0.39.0)"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["gql_book_server"]
